+++ MyProfile.Entity/Model/ToDoList.cs
@@ -9,7 +9,7 @@ namespace MyProfile.Entity.Model
 	{
 		[Key]
 		[DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-		public int ID { get; set; }
+		public long ID { get; set; }
 		public string Title { get; set; }
 		public DateTime DateCreate { get; set; }
 		public DateTime DateEdit { get; set; }
@@ -19,7 +19,7 @@ namespace MyProfile.Entity.Model
 		[ForeignKey("ToDoListFolder")]
 		public int ToDoListFolderID { get; set; }
 		[ForeignKey("VisibleElement")]
-		public int VisibleElementID { get; set; }
+		public long VisibleElementID { get; set; }
 
 		public virtual ToDoListFolder ToDoListFolder { get; set; }
 		public virtual VisibleElement VisibleElement { get; set; }
