+++ MyProfile.Entity/ModelView/Summary/SummaryFilter.cs
@@ -9,7 +9,7 @@ namespace MyProfile.Entity.ModelView
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public List<int>  SectionTypes { get; set; }
-        public List<int> Sections { get; set; }
+        public List<long> Sections { get; set; }
         public Guid UserID { get; set; }
         public PeriodTypesEnum PeriodType { get; set; }
     }
